refactored community sim to prepare for multiple gossipers
