+++ util/docker-compose.yml
@@ -12,5 +12,8 @@ services:
       MAX_NUM_PEERS: ${MAX_NUM_PEERS:-3}
       COMMUNITY_ID: ${COMMUNITY_ID:-c86a7db45eb3563ae047639817baec4db2bc7c25}
     volumes:
-      - ./torrents/:/app
+      - ./torrents/:/torrents
       - ./communitysim.py:/communitysim.py
+      - ./detoks_community.py:/detoks_community.py
+      - ./gossiper:/gossiper
+      - ./messages.py:/messages.py
